+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Weird Wares `0.2.2+1.21`
 ## Fixed
 - Don't play glitter sound when not used on a block.
+- Update glint across the network correctly.